move code_map::regmach into engine::regmach as code_map
This change will make it simpler to dissect stack machine engine implementation from register machine engine implementation.
