+++ crates/wasmi/src/engine/func_builder/regmach/result_mut.rs
@@ -19,7 +19,7 @@ use crate::{
             RegisterSpan,
             UnaryInstr,
         },
-        code_map::CompiledFuncEntity,
+        regmach::code_map::CompiledFuncEntity,
         CompiledFunc,
     },
     module::ModuleResources,
