+++ crates/wasmi/src/engine/mod.rs
@@ -24,7 +24,7 @@ use self::bytecode2::RegisterSpan;
 pub use self::{
     bytecode::DropKeep,
     bytecode2::Instruction as Instruction2,
-    code_map::{CodeMap2, CompiledFunc},
+    code_map::CompiledFunc,
     config::{Config, EngineBackend, FuelConsumptionMode},
     func_builder::{
         FuncBuilder,
@@ -34,6 +34,7 @@ pub use self::{
         RelativeDepth,
         TranslationError,
     },
+    regmach::CodeMap as CodeMap2,
     resumable::{ResumableCall, ResumableInvocation, TypedResumableCall, TypedResumableInvocation},
     stack::StackLimits,
     traits::{CallParams, CallResults},
@@ -42,12 +43,12 @@ use self::{
     bytecode::Instruction,
     bytecode2::{Provider, UntypedProvider},
     cache::InstanceCache,
-    code_map::{CodeMap, CompiledFuncEntity},
+    code_map::CodeMap,
     const_pool::{ConstPool, ConstPoolView, ConstRef},
     executor::{execute_wasm, WasmOutcome},
     func_builder::regmach::FuncLocalConstsIter,
     func_types::FuncTypeRegistry,
-    regmach::Stack as Stack2,
+    regmach::{code_map::CompiledFuncEntity, Stack as Stack2},
     resumable::ResumableCallBase,
     stack::{FuncFrame, Stack, ValueStack},
     trap::TaggedTrap,
