+++ crates/wasmi/src/engine/regmach/code_map.rs
