+++ crates/wasmi/src/engine/regmach/executor/instrs/store.rs
@@ -3,7 +3,7 @@ use crate::{
     core::{TrapCode, UntypedValue},
     engine::{
         bytecode2::{Const16, Instruction, Register, StoreAtInstr, StoreInstr, StoreOffset16Instr},
-        code_map::InstructionPtr2 as InstructionPtr,
+        regmach::code_map::InstructionPtr,
     },
 };
