+++ crates/wasmi/src/engine/regmach/stack/values.rs
@@ -1,7 +1,7 @@
 use super::err_stack_overflow;
 use crate::{
     core::UntypedValue,
-    engine::{bytecode2::Register, code_map::CompiledFuncEntity},
+    engine::{bytecode2::Register, regmach::code_map::CompiledFuncEntity},
 };
 use alloc::vec::Vec;
 use core::{fmt, fmt::Debug, iter, mem};
